Update alias.txt Add EDDK Sid Fix  Alias
There is still the issue with IniBuilds scenery not being updated in EDDK, resulting in aircraft using the default database not having SIDs for the 13/31 RWYs. There is a short link to the article in the knowledgebase, but a short alias for this issue would be great as Microsoft is not fixing the problem.
So the Alis would be helbfull for the next months until it got fixed.
But there is no "need" for it, but it would be a nice to have. :-)
Greetings Tobias